requirements.txt
# Core
anthropic==0.40.0
python-dotenv==1.0.0

# MCP
mcp==1.3.0

# Playwright
playwright==1.49.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# CLI
click==8.1.7

# Utilities
aiohttp==3.11.10
pydantic==2.10.3
